+++ apps/code-examples/src/app/features/forms.module.ts
@@ -3,11 +3,29 @@ import { RouterModule, Routes } from '@angular/router';
 
 import { CheckboxDemoComponent as CheckboxBasicDemoComponent } from '../code-examples/forms/checkbox/basic/checkbox-demo.component';
 import { CheckboxDemoModule as CheckboxBasicDemoModule } from '../code-examples/forms/checkbox/basic/checkbox-demo.module';
+import { CheckboxDemoComponent as InlineHelpCheckboxDemoComponent } from '../code-examples/forms/checkbox/inline-help/checkbox-demo.component';
+import { CheckboxDemoModule as InlineHelpCheckboxDemoModule } from '../code-examples/forms/checkbox/inline-help/checkbox-demo.module';
+import { RadioDemoComponent as InlineHelpRadioDemoComponent } from '../code-examples/forms/radio/inline-help/radio-demo.component';
+import { RadioDemoModule as InlineHelpRadioDemoModule } from '../code-examples/forms/radio/inline-help/radio-demo.module';
+import { ToggleSwitchDemoComponent as InlineHelpToggleSwitchComponent } from '../code-examples/forms/toggle-switch/inline-help/toggle-switch-demo.component';
+import { SkyToggleSwitchDemoModule as InlineHelpToggleSwitchModule } from '../code-examples/forms/toggle-switch/inline-help/toggle-switch-demo.module';
 
 const routes: Routes = [
   {
     path: 'checkbox/basic',
     component: CheckboxBasicDemoComponent,
+  },
+  {
+    path: 'radio/inline-help',
+    component: InlineHelpRadioDemoComponent,
+  },
+  {
+    path: 'checkbox/inline-help',
+    component: InlineHelpCheckboxDemoComponent,
+  },
+  {
+    path: 'toggle-switch/inline-help',
+    component: InlineHelpToggleSwitchComponent,
   },
 ];
 
@@ -18,6 +36,12 @@ const routes: Routes = [
 export class FormsRoutingModule {}
 
 @NgModule({
-  imports: [CheckboxBasicDemoModule, FormsRoutingModule],
+  imports: [
+    CheckboxBasicDemoModule,
+    FormsRoutingModule,
+    InlineHelpCheckboxDemoModule,
+    InlineHelpRadioDemoModule,
+    InlineHelpToggleSwitchModule,
+  ],
 })
 export class FormsModule {}