+++ libs/components/forms/src/lib/modules/toggle-switch/toggle-switch-label.component.ts
@@ -6,6 +6,7 @@ import { ChangeDetectionStrategy, Component } from '@angular/core';
 @Component({
   selector: 'sky-toggle-switch-label',
   templateUrl: './toggle-switch-label.component.html',
+  styleUrls: ['./toggle-switch-label.component.scss'],
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class SkyToggleSwitchLabelComponent {}